move getconfig into k8sutil
