+++ tools/vz/cmd/cmd_context.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2022, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package helpers
+package cmd
 
 import (
 	"io"
@@ -16,8 +16,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -37,7 +35,19 @@ func (rc *RootCmdContext) GetErrorStream() io.Writer {
 
 // GetClient - return a kubernetes client that supports the schemes used by the CLI
 func (rc *RootCmdContext) GetClient(cmd *cobra.Command) (client.Client, error) {
-	config, err := getConfig(cmd)
+	// Get command line value of kubeConfig location
+	kubeConfigLoc, err := cmd.Flags().GetString(constants.GlobalFlagKubeConfig)
+	if err != nil {
+		return nil, err
+	}
+
+	// Get command line value of kubeContext
+	context, err := cmd.Flags().GetString(constants.GlobalFlagContext)
+	if err != nil {
+		return nil, err
+	}
+
+	config, err := k8sutil.GetKubeConfigGivenPathAndContext(kubeConfigLoc, context)
 	if err != nil {
 		return nil, err
 	}
@@ -57,31 +67,4 @@ func NewRootCmdContext(streams genericclioptions.IOStreams) *RootCmdContext {
 	return &RootCmdContext{
 		IOStreams: streams,
 	}
-}
-
-// getConfig - create config attributes for Kubernetes client
-func getConfig(cmd *cobra.Command) (*rest.Config, error) {
-	// Determine the location of the kube config.  It was either specified on the command line
-	// or we find the default location.
-	kubeConfigLoc, err := cmd.Flags().GetString(constants.GlobalFlagKubeConfig)
-	if err != nil {
-		return nil, err
-	}
-	if len(kubeConfigLoc) == 0 {
-		kubeConfigLoc, err = k8sutil.GetKubeConfigLocation()
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	// Was a kube config context specified on the command line?
-	context, err := cmd.Flags().GetString(constants.GlobalFlagContext)
-	if err != nil {
-		return nil, err
-	}
-
-	// Create the config
-	return clientcmd.NewNonInteractiveDeferredLoadingClientConfig(
-		&clientcmd.ClientConfigLoadingRules{ExplicitPath: kubeConfigLoc},
-		&clientcmd.ConfigOverrides{CurrentContext: context}).ClientConfig()
 }
