+++ tools/vz/main.go
@@ -6,10 +6,11 @@ package main
 import (
 	"os"
 
+	"github.com/verrazzano/verrazzano/tools/vz/cmd"
+
 	"github.com/verrazzano/verrazzano/tools/vz/cmd/root"
 
 	"github.com/spf13/pflag"
-	"github.com/verrazzano/verrazzano/tools/vz/pkg/helpers"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 )
 
@@ -17,7 +18,7 @@ func main() {
 	flags := pflag.NewFlagSet("vz", pflag.ExitOnError)
 	pflag.CommandLine = flags
 
-	rc := helpers.NewRootCmdContext(genericclioptions.IOStreams{In: os.Stdin, Out: os.Stdout, ErrOut: os.Stderr})
+	rc := cmd.NewRootCmdContext(genericclioptions.IOStreams{In: os.Stdin, Out: os.Stdout, ErrOut: os.Stderr})
 	rootCmd := root.NewRootCmd(rc)
 	if err := rootCmd.Execute(); err != nil {
 		os.Exit(1)
